requirements.txt
cryptography==2.1.4
gunicorn==19.8.1
Flask_Ask==0.9.8
Flask==0.12.1
requests==2.19.1
psycopg2-binary==2.7.5
